+++ src/ground_station/flight_data/flight_data.js
@@ -51,6 +51,16 @@ module.exports = class FlightData {
       last_updated: "n/d",
     };
 
+    this.extension_data = {
+      last_updated_timestamp: "00:00",
+      last_updated_source: "tcp",
+      extensions: {
+        "test ext": {
+          status: "running",
+        },
+      },
+    };
+
     setInterval(this.#cycle.bind(this), FLIGHT_DATA_UPDATE_INTERVAL);
   }
 
@@ -63,6 +73,8 @@ module.exports = class FlightData {
       return this.phase_prediction;
     } else if (category === "error_frame") {
       return this.diagnostics.getGeneralData();
+    } else if (category === "extensions") {
+      return this.extension_data;
     } else {
       console.log("Error: Invalid category in FlightData.getData()");
     }
