+++ src/web_client/src/core/Nav.js
@@ -207,7 +207,6 @@ function NavBar({ navExpanded, setNavExpanded }) {
             path={"/ggs/command-center"}
           />
           <NavItem title="Telemetry" label="TLM" path={"/ggs/telemetry"} />
-          <NavItem title="Data" label="DAT" path={"/ggs/data-logs"} />
           <NavItem title="Diagnostic" label="DIA" path={"/ggs/diagnostics"} />
         </NavItem>
